refactor: src/ast, src/markdown
