+++ src/ast/transform-input-ast-to-output-commands.ts
@@ -1,11 +1,11 @@
 import { Nodes } from "npm:@types/mdast";
-import { astToMarkdown } from "../ast/ast-to-markdown.ts";
-import { extractFirstTopLevelHeadingString } from "../ast/extract-first-top-level-heading.ts";
-import { transformNode } from "../ast/transform-node.ts";
+import { astToMarkdown } from "./ast-to-markdown.ts";
+import { extractFirstTopLevelHeadingString } from "./extract-first-top-level-heading.ts";
+import { transformNode } from "./transform-node.ts";
 import { OutputCommand } from "../model/output-command.ts";
 import { ProjectId } from "../model/project-id.ts";
 
-export async function transformNodeToOutputCommands<
+export async function transformInputAstToOutputCommands<
   PI extends ProjectId = ProjectId,
 >(
   projectId: PI,
