+++ src/markdown/README.md
@@ -1,4 +1,4 @@
 # src/markdown
 
-Code that doesn't fit into the other categories, and that work with markdown
+Code that doesn't fit into the other categories, and that works with markdown
 source.
