+++ test/expect-input-to-output.ts
@@ -1,13 +1,13 @@
 import { sortBy } from "std/collections/sort_by.ts";
 import { assertEquals } from "std/assert/assert_equals.ts";
-import { transformAstToMarkdown } from "../mod.ts";
+import { transformInputAstToMarkdown } from "../mod.ts";
 import { markdownToAst } from "../src/ast/markdown-to-ast.ts";
 import { DeleteOrWriteFile } from "../src/model/output-command.ts";
 import { getMarkdownFilePathsInDirectory } from "../src/io/get-markdown-file-paths-in-directory.ts";
 import { readTextFilesToInputs } from "../src/io/read-text-files-to-inputs.ts";
-import { inputsToInputAsts } from "../src/markdown/inputs-to-input-asts.ts";
+import { inputsToInputAsts } from "../src/ast/inputs-to-input-asts.ts";
 
-import { transformAstsToOutputCommands } from "../src/markdown/transform-asts-to-output-commands.ts";
+import { transformInputAstsToOutputCommands } from "../src/ast/transform-input-asts-to-output-commands.ts";
 import { ProjectId } from "../src/model/project-id.ts";
 
 export function expectInputToOutput(
@@ -17,7 +17,7 @@ export function expectInputToOutput(
 ): () => Promise<void> {
   return async () => {
     const inputAst = markdownToAst(input.trim() + "\n");
-    const result = await transformAstToMarkdown(projectId, inputAst);
+    const result = await transformInputAstToMarkdown(projectId, inputAst);
     assertEquals(result, expectedOutput.trim() + "\n");
   };
 }
@@ -37,7 +37,7 @@ export function expectInputDirectoryToOutputs(
     const inputs = await readTextFilesToInputs(inputPaths);
     const inputAsts = inputsToInputAsts(inputs);
 
-    const actualOutputs = await transformAstsToOutputCommands(
+    const actualOutputs = await transformInputAstsToOutputCommands(
       projectId,
       inputAsts,
     );
